have a shared tlsconfig.go
